October v3 with location (#42)
* change version;

* upgrade to ocotber v3

* move files from cms to nocio headstart

* fix graphql

* - fixing schema
- fixing seeder
- fixing FilterResortsDirective

* fix unit test

* update readme and fixing error tags theme underflip

* OctoberV3

* Updated Continent Location Extension
Updated Resorts Seeder

* Fixed Tests

* Try changing x-debug config

* Try changing x-debug config

* Try changing x-debug config

* Try changing x-debug config

* Try changing env database

* Try changing env database

* Try changing env database

* Try changing env database

* Try changing env database

* Try changing env database

* Try changing env database

* Try adding in a migration

* Try adding in a migration

* Update docker-compose.yml.dist

* Try adding in a migration

* Try adding in a migration

* Try adding in a migration

* Try adding in a migration

* Try adding in a migration

* Try adding in a migration

* Try adding in a migration

* Try adding in a migration

* Try adding in a migration

* Try adding in a migration

* Try adding in a migration

* Try adding in a migration

* Try adding in a migration

* Try adding in a migration

* Try adding in a migration

* Try adding in a migration

* Try adding in a migration

* Try adding in a migration

* Try adding in a migration

* Try adding in a migration

* Try adding in a migration

* Try adding in a migration

* Try adding in a migration

* Try adding in a migration

* Try adding in a migration

* Try adding in a migration

* Make modified plugin directory for nocio

Make modified plugin directory for nocio

* Try adding in a migration

* Try adding in a migration

* Try adding in a migration

* Try adding in a migration

* Try adding in a migration

* Try adding in a migration

* Try adding in a migration

* Try adding in a migration

* Try adding in a migration

* Try adding in a migration

* Try adding in a migration

* Try adding in a migration

* Try adding in a migration

* Try adding in a migration

* Try adding in a migration

* Try modifying app key

* Try modifying depends_on:
                          - mysql

* add seed data to ci build

* Update config.yml

Make sure use new image

* Update docker-compose.yml.dist

Reorder container start and build

* Update config.yml

debugging circle ci test unit

* Update config.yml

Check environment

* Update config.yml

Update check environment

* Update config.yml

Debugging table resorts

* Update config.yml

add copy .env to circle ci

* Update config.yml

Clean debugging purpose of circleci

* circle ci config

* remove seed from circle ci config

* remove seed from circle ci config

circle ci config

Update config.yml

Clean debugging purpose of circleci

Update config.yml

add copy .env to circle ci

Update config.yml

Debugging table resorts

Update config.yml

Update check environment

Update config.yml

Check environment

Update config.yml

debugging circle ci test unit

Update docker-compose.yml.dist

Reorder container start and build

Update config.yml

Make sure use new image

add seed data to ci build

Try modifying depends_on:
                          - mysql

Try modifying app key

Try adding in a migration

Try adding in a migration

Try adding in a migration

Try adding in a migration

Try adding in a migration

Try adding in a migration

Try adding in a migration

Try adding in a migration

Try adding in a migration

Try adding in a migration

Try adding in a migration

Try adding in a migration

* Update README.md

Typo in README

* remove seed from circle ci config

* add coverage

* add coverage

* add coverage

* Update config.yml

Reorder ci jobs

* add coverage

* Update config.yml

Debuging tables migation

* Update config.yml

Debug DB table

* Update config.yml

Check DB tables

* Change phpunit.xml

* Add more tests

* Add more tests

* Add more tests

* Add more tests

---------

+++ modules/system/models/MailTemplate.php
@@ -7,7 +7,7 @@ use October\Rain\Mail\MailParser;
 use File as FileHelper;
 
 /**
- * Mail template
+ * MailTemplate
  *
  * @package october\system
  * @author Alexey Bobkov, Samuel Georges
@@ -17,28 +17,28 @@ class MailTemplate extends Model
     use \October\Rain\Database\Traits\Validation;
 
     /**
-     * @var string The database table used by the model.
+     * @var string table associated with the model
      */
     protected $table = 'system_mail_templates';
 
     /**
-     * @var array Guarded fields
+     * @var array guarded fields
      */
     protected $guarded = [];
 
     /**
-     * @var array Fillable fields
+     * @var array fillable fields
      */
     protected $fillable = [];
 
     /**
-     * @var array Validation rules
+     * @var array rules for validation
      */
     public $rules = [
-        'code'                  => 'required|unique:system_mail_templates',
-        'subject'               => 'required',
-        'description'           => 'required',
-        'content_html'          => 'required',
+        'code' => 'required|unique:system_mail_templates',
+        'subject' => 'required',
+        'description' => 'required',
+        'content_html' => 'required',
     ];
 
     public $belongsTo = [
@@ -46,41 +46,58 @@ class MailTemplate extends Model
     ];
 
     /**
-     * Returns an array of template codes and descriptions.
-     *
+     * listAllTemplates returns an array of template codes and descriptions.
      * @return array
      */
     public static function listAllTemplates()
     {
+        // Only return short codes, not views
         $fileTemplates = (array) MailManager::instance()->listRegisteredTemplates();
+        $fileTemplates = array_keys($fileTemplates);
+        $fileTemplates = array_combine($fileTemplates, $fileTemplates);
+
+        // Short codes are in the db too
         $dbTemplates = (array) self::lists('code', 'code');
         $templates = $fileTemplates + $dbTemplates;
+
+        // Sort and return
         ksort($templates);
         return $templates;
     }
 
     /**
-     * Returns a list of all mail templates.
-     *
-     * @return array Returns an array of the MailTemplate objects.
+     * allTemplates returns a list of all mail templates, returns an array of
+     * MailTemplate objects.
      */
-    public static function allTemplates()
+    public static function allTemplates(): array
     {
-        $result = [];
-        $codes = array_keys(self::listAllTemplates());
+        $dbTemplates = self::all();
+        $fileReferences = (array) MailManager::instance()->listRegisteredTemplates();
+        $fileTemplates = [];
 
-        foreach ($codes as $code) {
-            if (View::exists($code)) {
-                $result[] = self::findOrMakeTemplate($code);
+        // Add missing file-based references
+        foreach ($fileReferences as $code => $view) {
+            if ($dbTemplates->where('code', $code)->count() > 0) {
+                continue;
+            }
+
+            if (View::exists($view)) {
+                $template = new self;
+                $template->code = $code;
+                $template->fillFromView($view);
+                $fileTemplates[] = $template;
             }
         }
 
-        return $result;
+        if ($fileTemplates) {
+            $dbTemplates->merge($fileTemplates);
+        }
+
+        return $dbTemplates->all();
     }
 
     /**
-     * Syncronise all file templates to the database.
-     *
+     * syncAll synchronizes all file templates to the database.
      * @return void
      */
     public static function syncAll()
@@ -92,28 +109,20 @@ class MailTemplate extends Model
         $dbTemplates = self::lists('is_custom', 'code');
         $newTemplates = array_diff_key($templates, $dbTemplates);
 
-        /*
-         * Clean up non-customized templates
-         */
+        // Clean up non-customized templates
         foreach ($dbTemplates as $code => $isCustom) {
-            if ($isCustom) {
-                continue;
-            }
-
-            if (!array_key_exists($code, $templates)) {
+            if (!$isCustom && !array_key_exists($code, $templates)) {
                 self::whereCode($code)->delete();
             }
         }
 
-        /*
-         * Create new templates
-         */
-        foreach ($newTemplates as $code) {
-            $sections = self::getTemplateSections($code);
+        // Create new templates
+        foreach ($newTemplates as $code => $view) {
+            $sections = self::getTemplateSections($view);
             $layoutCode = array_get($sections, 'settings.layout', 'default');
             $description = array_get($sections, 'settings.description');
 
-            $template = self::make();
+            $template = new self;
             $template->code = $code;
             $template->description = $description;
             $template->is_custom = 0;
@@ -123,22 +132,18 @@ class MailTemplate extends Model
     }
 
     /**
-     * Fired after the model has been fetched.
-     *
-     * @return void
+     * afterFetch
      */
     public function afterFetch()
     {
         if (!$this->is_custom) {
-            $this->fillFromView($this->code);
+            $viewPath = MailManager::instance()->getViewPathForTemplate($this->code);
+            $this->fillFromView($viewPath ?: $this->code);
         }
     }
 
     /**
-     * Fill model using provided content.
-     *
-     * @param string $content
-     * @return void
+     * fillFromContent
      */
     public function fillFromContent($content)
     {
@@ -146,10 +151,7 @@ class MailTemplate extends Model
     }
 
     /**
-     * Fill model using a view file path.
-     *
-     * @param string $path
-     * @return void
+     * fillFromView
      */
     public function fillFromView($path)
     {
@@ -157,15 +159,12 @@ class MailTemplate extends Model
     }
 
     /**
-     * Fill model using provided section array.
-     *
-     * @param array $sections
-     * @return void
+     * fillFromSections
      */
     protected function fillFromSections($sections)
     {
-        $this->content_html = array_get($sections, 'html');
-        $this->content_text = array_get($sections, 'text');
+        $this->content_html = $sections['html'] ?? '';
+        $this->content_text = $sections['text'] ?? '';
         $this->subject = array_get($sections, 'settings.subject', 'No subject');
 
         $layoutCode = array_get($sections, 'settings.layout', 'default');
@@ -173,46 +172,48 @@ class MailTemplate extends Model
     }
 
     /**
-     * Get section array from a view file retrieved by code.
-     *
-     * @param string $code
-     * @return array|null
+     * getTemplateSections
      */
-    protected static function getTemplateSections($code)
+    protected static function getTemplateSections($view)
     {
-        if (!View::exists($code)) {
-            return null;
-        }
-        $view = View::make($code);
-        return MailParser::parse(FileHelper::get($view->getPath()));
+        return MailParser::parse(FileHelper::get(View::make($view)->getPath()));
     }
 
     /**
-     * Find a MailTemplate record by code or create one from a view file.
-     *
-     * @param string $code
-     * @return MailTemplate model
+     * findOrMakeTemplate
      */
     public static function findOrMakeTemplate($code)
     {
-        $template = self::whereCode($code)->first();
-
-        if (!$template && View::exists($code)) {
-            $template = new self;
-            $template->code = $code;
-            $template->fillFromView($code);
+        $template = self::where('code', $code)->first();
+        if ($template) {
+            return $template;
         }
 
-        return $template;
+        $view = MailManager::instance()->getViewPathForTemplate($code) ?: $code;
+        if (View::exists($view)) {
+            $template = new self;
+            $template->code = $code;
+            $template->fillFromView($view);
+            return $template;
+        }
+
+        return null;
     }
 
     /**
-     * @deprecated see System\Classes\MailManager::registerCallback
-     * Remove if year >= 2019
+     * canSendTemplate
      */
-    public static function registerCallback(callable $callback)
+    public static function canSendTemplate($code): bool
     {
-        traceLog('MailTemplate::registerCallback is deprecated, use ' . MailManager::class . '::registerCallback instead');
-        MailManager::instance()->registerCallback($callback);
+        if (!$code) {
+            return false;
+        }
+
+        $view = MailManager::instance()->getViewPathForTemplate($code) ?: $code;
+        if (View::exists($view)) {
+            return true;
+        }
+
+        return self::where('code', $code)->count() > 0;
     }
 }
